Salt turbo global hash with lockfile in self-hosted workspaces
Turbo hashes internal workspace dependencies by their files only, so
a lockfile-only bump to a plugin of an internally-vendored tool-config
package (e.g. an eslint plugin pinned via the root catalog) never
reaches sibling tasks that consume the plugin without a task edge.
CI then replays stale passing caches, as happened with the
eslint-plugin-unicorn v72 bump.

Emit pnpm-lock.yaml in the generated turbo.json globalDependencies
when discovery reports isSelfHosted. Consumers resolve tool-config
packages externally, where turbo's per-package transitive closure
already invalidates correctly, so their generated output is unchanged
and they keep granular caching.

Extract toTurboGlobs into turbo-globs.ts and split the
globalDependencies tests into their own file to stay within max-lines.
