+++ .changeset/turbo-lockfile-global-hash.md
@@ -1,0 +1,6 @@
+---
+---
+
+Salt the turbo global hash with `pnpm-lock.yaml` in self-hosted
+workspaces, where internally-vendored tool-config packages hide plugin
+bumps from sibling task hashes and CI replays stale passes
