+++ turbo.json
@@ -2,7 +2,8 @@
   "$schema": "https://turbo.build/schema.json",
   "globalDependencies": [
     "mise.lock",
-    "mise.toml"
+    "mise.toml",
+    "pnpm-lock.yaml"
   ],
   "tasks": {
     "//#lint:eslint": {
